finish

+++ src/controllers/user.ts
@@ -1,4 +1,4 @@
-import {asyncHandler, BadRequestError, deleteFile, Password, Request, Response, User} from "./controller";
+import {asyncHandler, BadRequestError, deleteFile, Password, Request, Response, Statistic, User} from "./controller";
 
 export const createUser = asyncHandler(async (req: Request, res: Response) => {
     const currentUser = await User.findOne({where: {login: req.body.login}});
@@ -11,6 +11,9 @@ export const createUser = asyncHandler(async (req: Request, res: Response) => {
 
     const user = await User.create(req.body);
     await user.save();
+    const statistic = await Statistic.findOne();
+    if(!statistic) await Statistic.create({users: 1});
+    statistic?.update({users: statistic.users + 1})
     res.status(201).json({
         status: 201,
         data: user
@@ -18,7 +21,17 @@ export const createUser = asyncHandler(async (req: Request, res: Response) => {
 })
 
 export const getAllUsers = asyncHandler(async (req: Request, res: Response) => {
-    const user = await User.findAll({attributes: {exclude: ['password']}});
+    const query = req.query;
+    const user = await User.findAll({where: query, attributes: {exclude: ['password']}});
+    res.status(200).json({
+        status: 200,
+        data: user
+    })
+})
+
+export const getUsersForAdmin = asyncHandler(async (req: Request, res: Response) => {
+    const query = req.query;
+    const user = await User.findAll({where: query,attributes: {exclude: ['password', 'login', 'image']}});
     res.status(200).json({
         status: 200,
         data: user
@@ -29,7 +42,7 @@ export const getAllUsers = asyncHandler(async (req: Request, res: Response) => {
 export const updateUserDetails = asyncHandler(
     async (req: Request, res: Response) => {
         const user = await User.findOne({where: {id: req.params.id}});
-        if (!user) throw new BadRequestError('Something went wrong.');
+        if (!user) throw new BadRequestError('User not found');
         if (req.body.password) {
             req.body.password = await Password.toHash(req.body.password);
         }
@@ -38,6 +51,8 @@ export const updateUserDetails = asyncHandler(
             if (user.image) deleteFile(user.image);
             req.body.image = req.file.path;
         }
+
+        if(req.body.role === 'superadmin') req.body.role=undefined;
 
         const updatedUser = await user.update(req.body);
         await updatedUser.save();
@@ -51,19 +66,21 @@ export const updateUserDetails = asyncHandler(
 
 export const updateUsersLoginAndPassword = asyncHandler(async (req: Request, res: Response) => {
     const user = await User.findOne({where: {id: req.user?.id}});
-    if (!user) throw new BadRequestError('Something went wrong.');
+    if (!user) throw new BadRequestError('User not found.');
     if (req.body.password) {
         if (!req.body.currentPassword) throw new BadRequestError("Please input current password.");
+        const password = user.password;
+        const currentPassword = req.body.currentPassword
         const matchPassword = await Password.compare(
-            user.password,
-            req.body.currentPassword,
+            password,
+            currentPassword,
         );
         if (matchPassword) {
             req.body.password = await Password.toHash(req.body.password);
             const updatedUser = await user.update({password: req.body.password});
             await updatedUser.save();
         } else {
-            throw new BadRequestError('Joriy parol xato kiritildi!');
+            throw new BadRequestError('Password is incorrect!');
         }
     }
 
@@ -95,8 +112,11 @@ export const updateUsersLoginAndPassword = asyncHandler(async (req: Request, res
 // Delete User
 export const deleteUser = asyncHandler(async (req: Request, res: Response) => {
     const user = await User.findOne({where: {id: req.params.id}});
+    if(user?.role === 'superadmin') throw new BadRequestError("You can't delete superadmin");
     if (!user) throw new BadRequestError("Something went wrong.");
     if (user.image) deleteFile(user.image);
+    const statistic = await Statistic.findOne();
+    statistic?.update({users: statistic.users - 1})
     await user.destroy();
     res.status(200).json({
         status: 200,
